refactor: `models` -> `types`
