+++ app/src/types/job.ts
